requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
requests==2.31.0
anthropic==0.34.0
python-dotenv==1.0.0
APScheduler==3.10.4
plotly==5.18.0
bcrypt==4.1.2